refactor: adjust import path for FinancialReporting component in App.ts
